Refactor feeder: multiple files
Move internal components to separate files.
